--- server/src/main/java/org/apache/iotdb/db/utils/TimeUtil.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.iotdb.db.utils;
-
-import java.util.concurrent.TimeUnit;
-
-public class TimeUtil {
-  public static TimeUnit toTimeUnit(String t) {
-    switch (t) {
-      case "h":
-        return TimeUnit.HOURS;
-      case "m":
-        return TimeUnit.MINUTES;
-      case "s":
-        return TimeUnit.SECONDS;
-      case "ms":
-        return TimeUnit.MILLISECONDS;
-      case "u":
-        return TimeUnit.MICROSECONDS;
-      case "n":
-        return TimeUnit.NANOSECONDS;
-      default:
-        throw new IllegalArgumentException("time precision must be one of: h,m,s,ms,u,n");
-    }
-  }
-}
